requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-dotenv>=1.0.0
playwright>=1.32.0
beautifulsoup4>=4.12.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
pydantic>=1.10.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'
